+++ scripts/iotop.py
@@ -38,7 +38,7 @@ except ImportError:
     sys.exit('platform not supported')
 
 import psutil
-from psutil._common import bytes2human
+from psutil import bytes2human
 
 win = curses.initscr()
 lineno = 0